fix(ci): resolve device-bridge quality-gate failures
- ruff: reformat bridge.py (join adjacent string literals).
- types-location: move exported bridge config types into config.types.ts.
- knip: un-export internally-only-used saveConfig and tokenizeCommand.
- syncpack: align bridge zod (^4.0.0) and @modelcontextprotocol/sdk (^1.27.1)
  with the rest of the workspace; refresh lockfile.
- interrogate: add one-line docstrings to device route/WS handlers and response
  schemas (75.9% -> 86.6%).
- bandit: nosec B105 on 'refresh_token: None' dict-field false positives.

Entire-Checkpoint: e73830665d14
