+++ apps/bridge/package.json
@@ -12,9 +12,9 @@
     "start": "node dist/index.js"
   },
   "dependencies": {
-    "@modelcontextprotocol/sdk": "^1.29.0",
+    "@modelcontextprotocol/sdk": "^1.27.1",
     "ws": "^8.18.0",
-    "zod": "^4.3.6"
+    "zod": "^4.0.0"
   },
   "devDependencies": {
     "@types/node": "^24.3.0",
